refactor: tidy up structure, resolve stray template dirs, relocate tests/styles, relax eslint configurations
