+++ __tests__/ai/aiDeclarationOverrideBug.test.ts
