--- hooks/useThemeColor.ts
@@ -1,23 +1,0 @@
-/**
- * Learn more about light and dark modes:
- * https://docs.expo.dev/guides/color-schemes/
- */
-
-import { Colors } from "../constants/Colors";
-import { useColorScheme } from "./useColorScheme";
-
-export function useThemeColor(
-  props: { light?: string; dark?: string },
-  colorName: keyof typeof Colors.light & keyof typeof Colors.dark,
-) {
-  const scheme = useColorScheme();
-  const theme: "light" | "dark" =
-    scheme === "dark" ? "dark" : "light";
-  const colorFromProps = props[theme];
-
-  if (colorFromProps) {
-    return colorFromProps;
-  } else {
-    return Colors[theme][colorName];
-  }
-}
